fix(nuxt): correct server side jwt logic
