+++ src/server/utils/networking.ts
@@ -33,19 +33,6 @@ export const useHost = ({ event }: { event?: H3Event } = {}) => {
   return host
 }
 
-export const getIsSecure = (
-  options:
-    | {
-        runtimeConfig: ReturnType<typeof useRuntimeConfig>
-      }
-    | {
-        siteUrl: URL
-      },
-) =>
-  ('siteUrl' in options
-    ? options.siteUrl.protocol
-    : getSiteUrl(options.runtimeConfig.public.i18n.baseUrl).siteUrlTyped
-        .protocol) === 'https:'
 export const useIsSecure = () =>
   getIsSecure({ runtimeConfig: useRuntimeConfig() })
