+++ src/shared/utils/networking.ts
@@ -10,6 +10,20 @@ export const getHost = (event: H3Event) => {
   return host
 }
 
+export const getIsSecure = (
+  options:
+    | {
+        runtimeConfig: ReturnType<typeof useRuntimeConfig>
+      }
+    | {
+        siteUrl: URL
+      },
+) =>
+  ('siteUrl' in options
+    ? options.siteUrl.protocol
+    : getSiteUrl(options.runtimeConfig.public.i18n.baseUrl).siteUrlTyped
+        .protocol) === 'https:'
+
 export const getServiceHref = ({
   host,
   isSsr = true,
